Move components for map interface into dedicated folder
