+++ src/app/app.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import {GraphDataManagerComponent} from "./graph-data-manager/graph-data-manager.component";
-import {TreeDataManagerComponent} from "./tree-data-manager/tree-data-manager.component";
+import {GraphDataManagerComponent} from "./app/graph-data-manager/graph-data-manager.component";
+import {TreeDataManagerComponent} from "./app/tree-data-manager/tree-data-manager.component";
 
 @Component({
   selector: 'app-root',
