+++ src/app/app/graph-data-detail/graph-data-detail.component.html
