+++ src/app/app/graph-data-display/graph-data-display.component.scss
