+++ src/app/app/graph-data-manager/graph-data-manager.component.scss
