+++ src/app/app/graph-data-selector/graph-data-selector.component.scss
